chore: tri par météo, optimisation
