+++ src/client/components/shared/Select.tsx
@@ -71,7 +71,7 @@ export const Select = Object.assign({}, SelectPrimitive, {
       <SelectPrimitive.Content
         className={clsxm(
           "bg-white rounded-md shadow-md border border-gray-100 z-50",
-          "!min-w-[300px] !max-w-[450px]",
+          "md:!min-w-[400px] !max-w-[450px]",
           "data-[state=open]:animate-dropdown-fade-in data-[state=closed]:animate-dropdown-fade-out",
           className,
         )}
